requirements.txt
streamlit==1.31.1
requests==2.31.0
pandas==2.1.4
numpy==1.26.3
scikit-learn==1.4.0
python-dotenv-1.2.1